[AOT] Refactor AOTExecutorCodegenModule using new AOT passes
This commit refactors the AOTExecutorCodegenModule to make use
of the newly introduced AOT passes AOTLowerMain,
CreateFunctionMetadata and CreateExecutor metadata. Some
modifications are additionally made to the 'Codegen' interface
to make important code generation options explicit.
